refactor: complete the record block state refactoring
